requirements.txt
absl-py==1.4.0
accelerate==0.24.1
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
batchgenerators==0.25
beautifulsoup4==4.11.2
bitsandbytes==0.37.0
bitsandbytes-cuda117==0.26.0.post2
bleach==6.0.0
blis==0.7.9
blobfile==2.0.1
bs4==0.0.1
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
cityscapesScripts==2.2.1
click==8.0.4
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
cmake==3.26.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
commonmark==0.9.1
confection==0.1.4
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
datasets==2.7.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.23.0
dill==0.3.6
distro==1.8.0
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.0
et-xmlfile==1.1.0
evaluate==0.4.1
executing==1.2.0
fastjsonschema==2.16.3
FastNLP==1.0.1
filelock==3.9.0
fire==0.5.0
fitlog==0.9.13
FlagEmbedding==1.1.6
Flask==2.2.3
flatbuffers==23.1.21
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.27.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
invisible-watermark==0.1.5
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==4.0.0
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-events==0.6.3
jupyter-nbextensions-configurator==0.6.1
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.10
langcodes==3.3.0
-e git+https://github.com/CompVis/stable-diffusion.git@21f890f9da3cfbeaba8e2ac3c425ee9e998d5229#egg=latent_diffusion
lazy_loader==0.2
lightning-utilities==0.7.1
linecache2==1.0.0
lit==15.0.7
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-collections==0.1.1
mpi4py==3.1.4
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.1.0
nltk==3.7
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
onnx==1.13.1
onnxruntime==1.14.1
openai==1.3.6
openai-api-call==0.3.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
peft @ file:///home/yu/ECNU/MELO-V2/peft-egg
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
preshed==3.0.8
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pycryptodomex==3.17
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pyquaternion==0.9.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.9.3
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.16.0
setproctitle==1.3.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.2.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.14.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.3
torchsummary==1.5.1
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.9.0
transformers==4.34.0
triton==2.0.0
typer==0.4.2
typing==3.7.4.3
typing-inspect==0.8.0
typing_extensions==4.8.0
unittest2==1.1.0
uri-template==1.2.0
urllib3==1.26.14
visdom==0.2.4
wandb==0.13.10
wasabi==0.10.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xformers==0.0.16
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
